moving sni to its own pkg
